refactor: move RecordExporter out of ExporterDirector
